TinyML_Topic3-main/TinyML_Topic3-main/requirements.txt
absl-py==2.3.1
altair==5.5.0
astunparse==1.6.3
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
click==8.3.0
contourpy==1.3.3
cycler==0.12.1
flatbuffers==25.9.23
fonttools==4.60.1
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
grpcio==1.75.1
h5py==3.14.0
idna==3.10
Jinja2==3.1.6
joblib==1.5.2
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==2.15.0
kiwisolver==1.4.9
libclang==18.1.1
Markdown==3.9
MarkupSafe==3.0.3
matplotlib==3.10.7
ml-dtypes==0.2.0
narwhals==2.7.0
numpy==1.26.4
oauthlib==3.3.1
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==4.25.8
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydeck==0.9.1
pyparsing==3.2.5
pyserial==3.5
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
six==1.17.0
smmap==5.0.2
streamlit==1.50.0
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
tensorflow-macos==2.15.0
tensorflow-metal==1.2.0
termcolor==3.1.0
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
wrapt==1.14.2
